Changes to Function 4 and Function4_Testing
Added:
1. Error Checking
If name is not in the sample set, will raise error exception

2. Remove unnecessary print statement
Since now function 4 is stable, its time to remove the print statement to show the true intended outcome.

3. Added test case
Added a test case to show that the error checking is working as intended. However not very happy that it will print out None after the error :(